added sever membership checks in chat and tickets

+++ backend/src/controllers/tickets.controller.js
@@ -1,21 +1,14 @@
 const Ticket = require("../models/tickets.model");
 const { validationResult } = require("express-validator");
-const jwt = require("jsonwebtoken");
-
-// Centralized error handling middleware
-const handleError = (res, error) => {
-    console.error(error);
-    res.status(500).json({ error: "Internal server error" });
-};
-
-const logEndPoint = (type, url) => {
-    console.log(new Date().toLocaleString(), "--->", type, " ", url);
-};
+const {
+    checkUserIsAdmin,
+    checkUserInServer,
+    logEndPoint,
+    handleError,
+} = require("../utils/util");
 
 // Create a new ticket
 exports.createTicket = async (req, res) => {
-    // Inside createTicket controller
-
     logEndPoint("POST", "/createTicket");
 
     const errors = validationResult(req);
@@ -27,6 +20,22 @@ exports.createTicket = async (req, res) => {
         const decoded = req.decoded;
         const { title, description, status, assigneeId, serverId } = req.body;
         const reporterId = decoded.id;
+
+        // Ensure the ticket is created in the by admin of server
+        const isAdmin = await checkUserIsAdmin(serverId, reporterId);
+        if (!isAdmin) {
+            return res
+                .status(403)
+                .json({ message: "You are not a admin in this server" });
+        }
+
+        // Ensure assignee is a member of the server
+        const isMember = await checkUserInServer(serverId, assigneeId);
+        if (!isMember) {
+            return res
+                .status(403)
+                .json({ message: "Assignee is not a member of this server" });
+        }
 
         const ticket = new Ticket({
             title,
@@ -47,11 +56,20 @@ exports.createTicket = async (req, res) => {
 
 // Get all tickets for the user's current server or a specified server
 exports.getAllTickets = async (req, res) => {
-    logEndPoint("GET", "/getAllTickets");
+    logEndPoint("GET", "/getAllTickets/:serverId");
 
     try {
-        // Get the serverId from the query parameters, if provided
         const queryServerId = req.params.serverId;
+        const decoded = req.decoded;
+        const userId = decoded.id;
+
+        // Ensure the ticket belongs to the user's server
+        const isMember = await checkUserInServer(queryServerId, userId);
+        if (!isMember) {
+            return res
+                .status(403)
+                .json({ message: "You are not a admin in this server" });
+        }
 
         const tickets = await Ticket.find({ server: queryServerId });
 
@@ -61,22 +79,22 @@ exports.getAllTickets = async (req, res) => {
     }
 };
 
-// Get a single ticket by ID
-exports.getTicketById = async (req, res) => {
-    logEndPoint("GET", "/getTicket/:ticketId");
+// // Get a single ticket by ID
+// exports.getTicketById = async (req, res) => {
+//     logEndPoint("GET", "/getTicket/:ticketId");
 
-    try {
-        const ticket = await Ticket.findById(req.params.ticketId);
+//     try {
+//         const ticket = await Ticket.findById(req.params.ticketId);
 
-        if (!ticket) {
-            return res.status(404).json({ error: "Ticket not found" });
-        }
+//         if (!ticket) {
+//             return res.status(404).json({ error: "Ticket not found" });
+//         }
 
-        res.json(ticket);
-    } catch (err) {
-        handleError(res, err);
-    }
-};
+//         res.json(ticket);
+//     } catch (err) {
+//         handleError(res, err);
+//     }
+// };
 
 // Update a ticket by ID
 exports.updateTicket = async (req, res) => {
@@ -90,6 +108,22 @@ exports.updateTicket = async (req, res) => {
     try {
         const { title, description, status, assigneeId } = req.body;
         const ticketId = req.params.ticketId;
+        const decoded = req.decoded;
+        const userId = decoded.id;
+
+        const ticket = await Ticket.findById(ticketId);
+        if (!ticket) {
+            return res.status(404).json({ error: "Ticket not found" });
+        }
+        const serverId = ticket.server;
+
+        // Ensure the ticket belongs to the user's server
+        const isAdmin = await checkUserIsAdmin(serverId, userId);
+        if (!isAdmin) {
+            return res
+                .status(403)
+                .json({ message: "You are not a admin in this server" });
+        }
 
         const updatedTicket = await Ticket.findByIdAndUpdate(
             ticketId,
@@ -122,6 +156,22 @@ exports.deleteTicket = async (req, res) => {
 
     try {
         const ticketId = req.params.ticketId;
+        const decoded = req.decoded;
+        const userId = decoded.id;
+
+        const ticket = await Ticket.findById(ticketId);
+        if (!ticket) {
+            return res.status(404).json({ error: "Ticket not found" });
+        }
+        const serverId = ticket.server;
+
+        // Ensure the ticket belongs to the user's server
+        const isAdmin = await checkUserIsAdmin(serverId, userId);
+        if (!isAdmin) {
+            return res
+                .status(403)
+                .json({ message: "You are not a admin in this server" });
+        }
 
         const deletedTicket = await Ticket.findByIdAndDelete(ticketId);
 
@@ -138,49 +188,49 @@ exports.deleteTicket = async (req, res) => {
     }
 };
 
-// Get tickets by status
-exports.getTicketsByStatus = async (req, res) => {
-    logEndPoint("GET", "/getTicketsByStatus");
+// // Get tickets by status
+// exports.getTicketsByStatus = async (req, res) => {
+//     logEndPoint("GET", "/getTicketsByStatus");
 
-    try {
-        const status = req.body.status;
-        const serverId = req.body.serverId;
-        //const token = req.headers['authorization'].split(' ')[1];
-        //const user = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
-        //const serverId = user.serverId;
-        console.log(status);
-        console.log(serverId);
-        const tickets = await Ticket.find({ server: serverId, status: status });
-        console.log("Reached here");
-        res.json(tickets);
-        console.log(tickets);
-    } catch (err) {
-        handleError(res, err);
-    }
-};
+//     try {
+//         const status = req.body.status;
+//         const serverId = req.body.serverId;
+//         //const token = req.headers['authorization'].split(' ')[1];
+//         //const user = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+//         //const serverId = user.serverId;
+//         console.log(status);
+//         console.log(serverId);
+//         const tickets = await Ticket.find({ server: serverId, status: status });
+//         console.log("Reached here");
+//         res.json(tickets);
+//         console.log(tickets);
+//     } catch (err) {
+//         handleError(res, err);
+//     }
+// };
 
-// Get tickets by assignee
-exports.getTicketsByAssignee = async (req, res) => {
-    logEndPoint("GET", "/getTicketsByAssignee");
+// // Get tickets by assignee
+// exports.getTicketsByAssignee = async (req, res) => {
+//     logEndPoint("GET", "/getTicketsByAssignee");
 
-    try {
-        const assigneeId = req.body.assigneeId;
-        const serverId = req.body.serverId;
+//     try {
+//         const assigneeId = req.body.assigneeId;
+//         const serverId = req.body.serverId;
 
-        //   const token = req.headers['authorization'].split(' ')[1];
-        //   const user = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
-        //   const serverId = user.serverId;
+//         //   const token = req.headers['authorization'].split(' ')[1];
+//         //   const user = jwt.verify(token, process.env.ACCESS_TOKEN_SECRET);
+//         //   const serverId = user.serverId;
 
-        const tickets = await Ticket.find({
-            server: serverId,
-            assignee: assigneeId,
-        });
+//         const tickets = await Ticket.find({
+//             server: serverId,
+//             assignee: assigneeId,
+//         });
 
-        res.json(tickets);
-    } catch (err) {
-        handleError(res, err);
-    }
-};
+//         res.json(tickets);
+//     } catch (err) {
+//         handleError(res, err);
+//     }
+// };
 
 // Change ticket status
 exports.changeTicketStatus = async (req, res) => {
@@ -189,11 +239,22 @@ exports.changeTicketStatus = async (req, res) => {
     try {
         const status = req.body.status;
         const ticketId = req.body.ticketId;
+        const decoded = req.decoded;
+        const userId = decoded.id;
 
         // Ensure the ticket belongs to the user's server
-        const ticket = await Ticket.findOne({ _id: ticketId });
+        const ticket = await Ticket.findById(ticketId);
         if (!ticket) {
             return res.status(404).json({ error: "Ticket not found" });
+        }
+        const assigneeId = ticket.assignee;
+
+        const isAdmin = await checkUserIsAdmin(ticket.server, userId);
+        if (!isAdmin && userId !== assigneeId) {
+            return res.status(403).json({
+                message:
+                    "You need to be an admin in this server or assignee of the ticket",
+            });
         }
 
         ticket.status = status;
